verify_types.rs: extract
